refactor : domain 패키지가 커짐에 따라 패키지 구분
