+++ src/main/java/LadderGame.java
@@ -1,7 +1,7 @@
-import domain.Height;
-import domain.Ladder;
-import domain.Names;
-import domain.bridgeConstructstrategy.RandomBridgeConstructStrategy;
+import domain.ladder.Height;
+import domain.ladder.Ladder;
+import domain.player.Names;
+import domain.ladder.bridgeConstructstrategy.RandomBridgeConstructStrategy;
 import view.InputView;
 import view.OutputView;
