+++ src/main/java/domain/ladder/Bridges.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.ladder;
 
 import java.util.Collections;
 import java.util.List;
